refactor: refactoring all code(holy shmoly jesus god damn it)
