--- data/all_data.py
@@ -1,101 +1,0 @@
-# TIMETABLE = {
-#     "Monday": {
-#         "08:30": "네트워크 구축",
-#         "09:30": "네트워크 구축",
-#         "10:30": "네트워크 구축",
-#         "11:30": "네크워크 구축",
-#         "12:20": "점심 시간",
-#         "13:10": "일본어",
-#         "14:10": "미적분"
-#     },
-    
-#     "Tuesday": {
-#         "08:30": "자료구조",
-#         "09:30": "자료구조",
-#         "10:30": "자료구조",
-#         "11:30": "일본어",
-#         "12:20": "점심 시간",
-#         "13:10": "성공적인 직업생활",
-#         "14:10": "응용 프로그래밍",
-#         "15:10": "응용 프로그래밍"
-#     },
-    
-#     "Wednesday": {
-#         "08:30": "스포츠",
-#         "09:30": "진로",
-#         "10:30": "성공적인 직업생활",
-#         "11:30": "응용 프로그래밍",
-#         "12:20": "점심 시간",
-#         "13:10": "응용 프로그래밍",
-#         "14:10": "미적분"
-#     },
-    
-#     "Thursday": {
-#         "08:30": "클라우드 보안",
-#         "09:30": "클라우드 보안",
-#         "10:30": "클라우드 보안",
-#         "11:30": "미적분",
-#         "12:20": "점심 시간",
-#         "13:10": "성공적인 직업생활",
-#         "14:10": "일본어",
-#         "15:10": "일본어",
-#     },
-    
-#     "Friday": {
-#         "08:30": "빅데이터 분석",
-#         "09:30": "빅데이터 분석",
-#         "10:30": "빅데이터 분석",
-#         "11:30": "미적분",
-#         "12:20": "점심 시간",
-#         "13:10": "음악",
-#         "14:10": "자율",
-        
-#     },
-# }
-
-SHORTENED_TIMETABLE = {
-    "MWF": {
-        "08:40": "1교시",
-        "09:30": "2교시",
-        "10:20": "3교시",
-        "11:10": "4교시",
-        "11:40": "점심 시간",
-        "12:40": "5교시",
-        "13:30": "6교시",
-    },
-    
-    "TT": {
-        "08:40": "1교시",
-        "09:30": "2교시",
-        "10:20": "3교시",
-        "11:10": "4교시",
-        "11:40": "점심 시간",
-        "12:40": "5교시",
-        "13:30": "6교시",
-        "14:20": "7교시",
-    }
-    
-}
-
-BREAKTIME = {
-    "MWF": {
-        "09:30": "2교시",
-        "10:30": "3교시",
-        "11:30": "4교시",
-        "14:10": "5교시",
-        "15:10": "6교시",
-    },
-    
-    "TT": {
-        "09:30": "2교시",
-        "10:30": "3교시",
-        "11:30": "4교시",
-        "12:20": "점심 시간",
-        "14:10": "5교시",
-        "15:10": "6교시",
-        "16:10": "7교시"
-    }
-        
-}
-
-BIRTHDAY = "03-09"
