+++ data/breaktime.json
@@ -1,0 +1,19 @@
+{
+    "MWF": {
+        "09:30": "2교시",
+        "10:30": "3교시",
+        "11:30": "4교시",
+        "14:10": "5교시",
+        "15:10": "6교시"
+    },
+
+    "TT": {
+        "09:30": "2교시",
+        "10:30": "3교시",
+        "11:30": "4교시",
+        "12:20": "점심 시간",
+        "14:10": "5교시",
+        "15:10": "6교시",
+        "16:10": "7교시"
+    }
+}
