+++ data/shortenedTimetable.json
@@ -1,0 +1,22 @@
+{
+    "MWF": {
+        "08:40": "1교시",
+        "09:30": "2교시",
+        "10:20": "3교시",
+        "11:10": "4교시",
+        "11:40": "점심 시간",
+        "12:40": "5교시",
+        "13:30": "6교시"
+    },
+
+    "TT": {
+        "08:40": "1교시",
+        "09:30": "2교시",
+        "10:20": "3교시",
+        "11:10": "4교시",
+        "11:40": "점심 시간",
+        "12:40": "5교시",
+        "13:30": "6교시",
+        "14:20": "7교시"
+    }
+}
