+++ data/timetable.json
@@ -27,7 +27,7 @@
         "12:20": "점심 시간",
         "13:10": "응용 프로그래밍",
         "14:10": "미적분",
-        "15:10": "ㅎㅇ"
+        "15:10": ""
     },
     "Thursday": {
         "08:30": "클라우드 보안",
@@ -37,7 +37,7 @@
         "12:20": "점심 시간",
         "13:10": "성공적인 직업생활",
         "14:10": "일본어",
-        "15:10": "일본어"
+        "15:10": "일본어2"
     },
     "Friday": {
         "08:30": "빅데이터 분석",
