--- function/isMWF.py
@@ -1,13 +1,0 @@
-def isMWF(today:str):
-    """오늘이 월수금 인지 확인해주는 함수
-
-    Args:
-        today (str): 오늘 요일을 받아옴
-
-    Returns:
-        bool: 오늘이 월수금이면 True를 아니면 False를 반환
-    """
-    if today in ["Monday", "Wednesday", "Friday"]:
-        return True
-    else:
-        return False
