+++ function/systemTray.py
@@ -1,12 +1,8 @@
 import sys, logging, json, os
-from function.filePath import assets_dir_func, data_dir_func
 from PyQt5.QtWidgets import QApplication, QSystemTrayIcon, QMenu, QAction
 from PyQt5.QtGui import QIcon
-from data.all_data import SHORTENED_TIMETABLE
-from function.todayVariable import todayVariable
-from function.isMWF import isMWF
-from function.isShortened import isShortened
-from function.settings import settings
+from function.functions import todayVariable, isMWF, isShortened, assets_dir_func, data_dir_func
+from function.systemTrayFunc import settings
 
 isActivated = True
 class systemTray:
@@ -44,12 +40,17 @@ class systemTray:
 
     def update_tooltip(self):
         # 저장할 파일 경로
-        FILE_PATH = data_dir_func("timetable.json")
+        TIMETABLE_PATH = data_dir_func("timetable.json")
+        SHORTENED_TIMETABLE_PATH = data_dir_func("shortenedTimetable.json")
 
         # 기존 시간표 데이터 불러오기
-        if os.path.exists(FILE_PATH):
-            with open(FILE_PATH, "r", encoding="utf-8") as f:
-                TIMETABLE = json.load(f)
+        os.path.exists(TIMETABLE_PATH)
+        os.path.exists(SHORTENED_TIMETABLE_PATH)
+        
+        with open(TIMETABLE_PATH, "r", encoding="utf-8") as f:
+            TIMETABLE = json.load(f)
+        with open(SHORTENED_TIMETABLE_PATH, "r", encoding='utf-8') as f:
+            SHORTENED_TIMETABLE = json.load(f)
         
         global isActivated
